# Conflicts:
#	go.mod
#	go.sum
#	internal/services/containerapps/client/client.go
#	internal/services/containerapps/container_app_environment_certificate_resource.go

+++ internal/services/sentinel/sentinel_alert_rule_anomaly_built_in_resource.go
@@ -28,7 +28,7 @@ type AlertRuleAnomalyBuiltInModel struct {
 	Enabled                      bool                                    `tfschema:"enabled"`
 	Mode                         string                                  `tfschema:"mode"`
 	AnomalyVersion               string                                  `tfschema:"anomaly_version"`
-	AnomalySettingsVersion       int32                                   `tfschema:"anomaly_settings_version"`
+	AnomalySettingsVersion       int64                                   `tfschema:"anomaly_settings_version"`
 	Description                  string                                  `tfschema:"description"`
 	Frequency                    string                                  `tfschema:"frequency"`
 	RequiredDataConnectors       []AnomalyRuleRequiredDataConnectorModel `tfschema:"required_data_connector"`
@@ -288,7 +288,7 @@ func (r AlertRuleAnomalyBuiltInResource) Read() sdk.ResourceFunc {
 			}
 
 			if resp.AnomalySettingsVersion != nil {
-				state.AnomalySettingsVersion = *resp.AnomalySettingsVersion
+				state.AnomalySettingsVersion = int64(*resp.AnomalySettingsVersion)
 			}
 
 			if resp.Description != nil {
